requirements.txt
requests>=2.31.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
